requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow>=10.0.1
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.7
pydantic>=2.5.0
python-multipart>=0.0.6
opencv-python>=4.8.0
numpy>=1.24.0
